+++ Framework/PythonInterface/mantid/utils/__init__.py
@@ -4,4 +4,4 @@
 #   NScD Oak Ridge National Laboratory, European Spallation Source,
 #   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 # SPDX - License - Identifier: GPL - 3.0 +
-from .utils import *
+from ._utils import *
