+++ Framework/PythonInterface/mantid/utils/_utils.py
